Allow building with CMake 4.0
Bump cmake_minumum_required to 3.10, which matches what most other modules
in libyui use (3.5 is the minimum required to be supportable by CMake 4.0)

Remove CMP0046, as CMake 4.0 no longer supports it. The OLD mode in any case
was a workaround for 'badly written CMakeLists.txt', as it just instructed
cmake to ignore non-existing dependenices listed in add_dependencies.
This package does not rely on non-existing dependencies and builds perfectly
without this policy.